+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## Enhancements
 * `transform_ilr()` gained a new argument allowing the specify the base of the transformation.
+* `barplot()` gained a new argument to display a legend.
 
 # nexus 0.1.0